+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJWA1FZMC280CP6K9HKZJK41"
+        "01KJWE7VSKP7DRN9JHE034GVBR"
     )
     r"""The unique identifier of the data source"""