+++ packages/orq-rc/RELEASES.md
@@ -6299,3 +6299,13 @@ Based on:
 - [python v4.5.0-rc.25] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.25 - packages/orq-rc
+
+## 2026-03-04 12:50:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.740.1 (2.846.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.26] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.26 - packages/orq-rc